Converting the FileConversion code to use the newer 'pipeline' methodology which required the converion of all of the writers to inherit from AbstractFilter. The H5VoxelGrainIdReader is not complete at this point

+++ Source/DREAM3DLib/ReconstructionFilters/CleanupGrains.cpp
@@ -134,7 +134,7 @@ void CleanupGrains::execute()
   }
   int numgrains = m->m_Grains.size();
   neighborlist.resize(numgrains);
-  for(size_t i=0;i<numgrains;i++)
+  for(int i=0;i<numgrains;i++)
   {
 	  neighborlist[i] = find_neighbors->neighborlist[i];
   }
